refactor: move BoardCardModal to HomePage
